Add integration tests for user service and hiring manager routes
- Implement integration tests for user service functionality, including user registration and role updates.
- Create tests for hiring manager route functionality, ensuring proper access control and redirection.
- Add tests for index dashboard selection and routing flow, covering employee and hiring manager scenarios.
- Introduce privacy consent utility tests to validate registration and consent handling.
- Ensure French locale support in routing and registration flows.
